+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 ## 1.2.2
 
 - [FIX] "View TV Settings" causes HomeBridge crash #207 (if no current input)
+- [FIX] Input set in a scene is ignored, and main activity is turned on #157
 
 ## 1.2.1